rename Vid_to_Strain to arc_length_strain behind a compatibility shim
